trying to get models and app connected
