+++ app/models.py
@@ -1,4 +1,4 @@
-from app import db
+from flask_app import db
 
 class Locations(db.Model):
